Fix "Undefined-shift in sbr_envelope"
For some huffman tables result is left-shifted by 1 bit (delta).
When value is negative, it is UB in C (in C++ that would be OK).

Desired result of left-shifting is multiplication by 2.
It happens that tables whose values are left-shifted have small
value range. Thus it is possible to "bake" this multiplication into
the Huffman tables.

Formulae that transforms table value to result is `x + 64`.
To multiply encoded values we do decode-multiply-encode:
`(x + 64) * 2 - 64 == x * 2 + 64`.
To check that values are still valid, we ensure those are
in the range -128..-1.

This is applied only to negative values of:
 - `t_huffman_env_bal_1_5dB`
 - `f_huffman_env_bal_1_5dB`
 - `t_huffman_env_bal_3_0dB`
 - `f_huffman_env_bal_3_0dB`
 - `t_huffman_noise_bal_3_0dB`