Fix UB in fixed_point TNS
Problem: malicuous / corrupted input might produce very large TNS
filter coefficients; in regular files that does not happen,
because such filters make no sense.

Values produced by specially crafted filter might reach almost 2**20.
Of course such filter is quite unstable.
For floating-point processing that is not problem (besides garbage
output). For fixed point it causes UB, because "coef" number type
has range [-8.0,8.0).

This changelist changes how TNS coefficients are calculated and
applied in fixed_point build. Floating-point behaviour is not
changed - (un)scaling is not applied, "exp" is ignored.

After each iteration of TNS filter calculation it is checked,
if the sum of abs values of filter is not greater than 3.5;
if check failed, then filter values are divided by 2 and an
"exp" return value is incremented.

If "exp" becomes too large, filter is not applied (since it
renders to be harmful for sample values).

Otherwise "exp" is used to scale the result of convolution
with filter.