Fix integer overflow in estimate_envelope
We accumulating sum of squares. With usual REAL values it limits us to source values below 256.
That is not much (values of 10k sometimes appear in the stream).
To fix that we exchance some precision and when squaring REAL values we use MUL_C.
As a result squares are prescaled by 2^-14. I've modified range check to limit
output to maximal REAL range after division and un-prescaling.